refactor(ui-select): extract styles and tokens into ui-select.styles.ts
Split 1192-line file into ui-select.ts (717 lines, component class) and
ui-select.styles.ts (474 lines, tokens + CSS). No behavioral changes.
